requirements.txt
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-telegram-bot==22.1
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typing_extensions==4.14.0
urllib3==2.5.0
